requirements.txt
python-dotenv
langchain
langchainhub
langgraph
langchain_openai
playwright
ipython